+++ content-cache/tests/integration/conftest.py
@@ -21,6 +21,7 @@ logger = logging.getLogger(__name__)
 
 CONFIG_CHARM_NAME = "content-cache-backends-config"
 CERT_CHARM_NAME = "self-signed-certificates"
+METRIC_CHARM_NAME = "grafana-agent"
 
 
 @pytest.fixture(name="app_name", scope="module")
@@ -35,10 +36,22 @@ def config_app_name_fixture() -> str:
     return "config"
 
 
+@pytest.fixture(name="config_alt_app_name", scope="module")
+def config_alt_app_name_fixture() -> str:
+    """The application name for the alternative configuration charm."""
+    return "config-alt"
+
+
 @pytest.fixture(name="cert_app_name", scope="module")
 def cert_app_name_fixture() -> str:
     """The application name for the TLS certificate charm."""
     return "cert"
+
+
+@pytest.fixture(name="metric_app_name", scope="module")
+def metric_app_name_fixture() -> str:
+    """The application name of the metric export charm."""
+    return "metric"
 
 
 @pytest.fixture(name="charm_file", scope="module")
@@ -50,8 +63,15 @@ def charm_file_fixture(pytestconfig: pytest.Config) -> str:
 
 
 @pytest_asyncio.fixture(name="config_charm_file", scope="module")
-async def config_charm_file_fixture(ops_test: OpsTest) -> AsyncIterator[str]:
+async def config_charm_file_fixture(
+    ops_test: OpsTest, pytestconfig: pytest.Config
+) -> AsyncIterator[str]:
     """Build the configuration charm file and return the path."""
+    file = pytestconfig.getoption("--config-charm-file")
+    if file:
+        yield file
+        return
+
     path = await ops_test.build_charm("../content-cache-backends-config")
     yield str(path)
 
@@ -69,18 +89,59 @@ async def deploy_applications_fixture(
     config_charm_file: str,
     app_name: str,
     config_app_name: str,
+    config_alt_app_name: str,
     cert_app_name: str,
+    metric_app_name: str,
+    pytestconfig: pytest.Config,
 ) -> AsyncIterator[dict[str, Application]]:
     """Deploy all applications in parallel."""
-    app_task = model.deploy(charm_file, app_name, base="ubuntu@24.04")
-    config_app_task = model.deploy(config_charm_file, config_app_name, num_units=0)
-    cert_app_task = model.deploy(
-        CERT_CHARM_NAME, cert_app_name, base="ubuntu@22.04", channel="latest/edge"
+    if pytestconfig.getoption("--no-deploy"):
+        try:
+            res = {
+                app_name: model.applications[app_name],
+                config_app_name: model.applications[config_app_name],
+                config_alt_app_name: model.applications[config_alt_app_name],
+                cert_app_name: model.applications[cert_app_name],
+                metric_app_name: model.applications[metric_app_name],
+            }
+        except KeyError:
+            raise RuntimeError("At least one app is missing, you cannot use --no-deploy.")
+        yield res
+        return
+
+    app_deploy = model.deploy(charm_file, app_name, base="ubuntu@24.04")
+    config_app_deploy = model.deploy(
+        config_charm_file, config_app_name, base="ubuntu@24.04", num_units=0
     )
-    app, config_app, cert_app = await asyncio.gather(app_task, config_app_task, cert_app_task)
+    config_alt_app_deploy = model.deploy(
+        config_charm_file, config_alt_app_name, base="ubuntu@24.04", num_units=0
+    )
+    cert_app_deploy = model.deploy(
+        CERT_CHARM_NAME, cert_app_name, channel="latest/edge", base="ubuntu@22.04"
+    )
+    # The pinning to revision 319 due to a `model.deploy` issue. Ideally, the revision is not
+    # pinned. The `model.deploy` is unable to resolve to a workable revision, hence hardcoding to
+    # revision 319.
+    metric_app_deploy = model.deploy(
+        METRIC_CHARM_NAME,
+        metric_app_name,
+        channel="latest/edge",
+        base="ubuntu@24.04",
+        revision=319,
+        num_units=0,
+    )
+    app, config_app, config_alt_app, cert_app, metric_app = await asyncio.gather(
+        app_deploy, config_app_deploy, config_alt_app_deploy, cert_app_deploy, metric_app_deploy
+    )
     await model.wait_for_idle([app.name], status="blocked", timeout=15 * 60)
     await model.wait_for_idle([cert_app.name], status="active", timeout=15 * 60)
-    yield {app_name: app, config_app_name: config_app, cert_app_name: cert_app}
+    yield {
+        app_name: app,
+        config_app_name: config_app,
+        config_alt_app_name: config_alt_app,
+        cert_app_name: cert_app,
+        metric_app_name: metric_app,
+    }
 
 
 @pytest_asyncio.fixture(name="app", scope="module")
@@ -99,12 +160,28 @@ async def config_app_fixture(
     yield applications[config_app_name]
 
 
+@pytest_asyncio.fixture(name="config_alt_app", scope="module")
+async def config_alt_app_fixture(
+    config_alt_app_name: str, applications: dict[str, Application]
+) -> AsyncIterator[Application]:
+    """The alternative configuration charm application for testing."""
+    yield applications[config_alt_app_name]
+
+
 @pytest_asyncio.fixture(name="cert_app", scope="module")
 async def cert_app_fixture(
     cert_app_name: str, applications: dict[str, Application]
 ) -> AsyncIterator[Application]:
     """The TLS certificate charm application for testing."""
     yield applications[cert_app_name]
+
+
+@pytest_asyncio.fixture(name="metric_app", scope="module")
+async def metric_app_fixture(
+    metric_app_name: str, applications: dict[str, Application]
+) -> AsyncIterator[Application]:
+    """The metric agent charm application for testing."""
+    yield applications[metric_app_name]
 
 
 @pytest.fixture(name="http_ok_path", scope="module")
@@ -140,11 +217,15 @@ async def http_ok_ip_fixture(http_ok_app: Application) -> str:
 
 @pytest_asyncio.fixture(name="cache_tester", scope="function")
 async def cache_tester_fixture(
-    model: Model, app: Application, config_app: Application, cert_app: Application
+    model: Model,
+    app: Application,
+    config_app: Application,
+    config_alt_app: Application,
+    cert_app: Application,
 ) -> AsyncIterator[CacheTester]:
     """Get the cache tester."""
     unit = app.units[0]
-    tester = CacheTester(model, app, config_app, cert_app)
+    tester = CacheTester(model, app, config_app, config_alt_app, cert_app)
 
     yield tester